Check that AZs are balanced after an ASG is scaled
AWS does not guarantee that nodes will be balanced across AZ when
scaling up (which can happen, for example, if one zone does not have
enough capacity at the time). For an AWS cluster that uses EBS volumes,
this can be problematic and cause Pods to be unschedulable when one zone
has reduced capacity. This check will abort draining of the nodes if it
detects that the nodes are not evenly spread across three AZs.

Signed-off-by: Dimitrios Karagiannis <dhkarag@gmail.com>