Do not suspend the ReplaceUnhealthy process
The AWS documentation mentions that when the Launch process is
suspended, "ReplaceUnhealthy continues to terminate unhealthy instances,
but does not launch replacements". We actually terminate instances
manually, through this script but the ReplaceUnhealthy process allows
the ASG to acknowledge these terminated instances. It actually does a
no-op termination itself before removing them from the ASG pool.

With this change, we should now resize the ASG _after_ the instance
count has reached the desired number which should result in less
turbulence after resuming the previously suspended processes.

Signed-off-by: Dimitrios Karagiannis <dhkarag@gmail.com>