Drain nodes in batches to avoid large scale outs
Doubling the size of the ASG is only practical up to a limit. That, in
addition with the likelihood of bumping onto AWS capacity issues
dictates a different solution.

With this change, we aim to drain nodes in batches (of three, by
default). This means that we only need to scale out by a number of nodes
equal to the "batch size" and allow the ASG to replace terminated
nodes with new ones as we go.

Signed-off-by: Dimitrios Karagiannis <dhkarag@gmail.com>