Fix last batch condition when resuming nodes
`nc` tracks the amount of ready nodes, and it should be compared with
the amount of ready nodes we ultimately want to have (asg_count).
Checking against `old_nodes_count` worked on initial rolls because the
amount of old nodes at startup was the same as the asg_count, but breaks
on resumed rolls.